+++ external-crates/move/crates/move-compiler/src/diagnostics/codes.rs
@@ -6,8 +6,6 @@
 // Main types
 //**************************************************************************************************
 
-use crate::shared::FILTER_ALL;
-
 #[derive(PartialEq, Eq, Clone, Copy, Debug, Hash, PartialOrd, Ord)]
 pub enum Severity {
     Note = 0,
@@ -20,8 +18,6 @@ pub enum Severity {
 /// A an optional prefix to distinguish between different types of warnings (internal vs. possibly
 /// multiple externally provided ones).
 pub type ExternalPrefix = Option<&'static str>;
-/// The name for a well-known filter.
-pub type WellKnownFilterName = &'static str;
 /// The ID for a diagnostic, consisting of an optional prefix, a category, and a code.
 pub type DiagnosticsID = (ExternalPrefix, u8, u8);
 
@@ -53,26 +49,6 @@ pub(crate) trait DiagnosticCode: Copy {
             message,
         }
     }
-}
-
-#[derive(PartialEq, Eq, Clone, Copy, Debug, PartialOrd, Ord)]
-/// Represents a single annotation for a diagnostic filter
-pub enum WarningFilter {
-    /// Filters all warnings
-    All(ExternalPrefix),
-    /// Filters all warnings of a specific category. Only known filters have names.
-    Category {
-        prefix: ExternalPrefix,
-        category: u8,
-        name: Option<WellKnownFilterName>,
-    },
-    /// Filters a single warning, as defined by codes below. Only known filters have names.
-    Code {
-        prefix: ExternalPrefix,
-        category: u8,
-        code: u8,
-        name: Option<WellKnownFilterName>,
-    },
 }
 
 //**************************************************************************************************
@@ -386,45 +362,6 @@ codes!(
         PathAutocomplete: { msg: "IDE path autocomplete", severity: Note },
     ],
 );
-
-//**************************************************************************************************
-// Warning Filter
-//**************************************************************************************************
-
-impl WarningFilter {
-    pub fn to_str(self) -> Option<&'static str> {
-        match self {
-            Self::All(_) => Some(FILTER_ALL),
-            Self::Category { name, .. } | Self::Code { name, .. } => name,
-        }
-    }
-
-    pub fn code(
-        prefix: ExternalPrefix,
-        category: u8,
-        code: u8,
-        name: Option<WellKnownFilterName>,
-    ) -> Self {
-        Self::Code {
-            prefix,
-            category,
-            code,
-            name,
-        }
-    }
-
-    pub fn category(
-        prefix: ExternalPrefix,
-        category: u8,
-        name: Option<WellKnownFilterName>,
-    ) -> Self {
-        Self::Category {
-            prefix,
-            category,
-            name,
-        }
-    }
-}
 
 //**************************************************************************************************
 // impls
