+++ external-crates/move/crates/move-compiler/src/diagnostics/mod.rs
@@ -3,19 +3,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod codes;
+pub mod warning_filters;
 
 use crate::{
     command_line::COLOR_MODE_ENV_VAR,
-    diagnostics::codes::{
-        Category, DiagnosticCode, DiagnosticInfo, ExternalPrefix, Severity, WarningFilter,
-        WellKnownFilterName,
-    },
-    shared::{
-        ast_debug::AstDebug,
-        files::{ByteSpan, FileByteSpan, FileId, MappedFiles},
-        known_attributes, FILTER_UNUSED_CONST, FILTER_UNUSED_FUNCTION, FILTER_UNUSED_MUT_PARAM,
-        FILTER_UNUSED_MUT_REF, FILTER_UNUSED_STRUCT_FIELD, FILTER_UNUSED_TYPE_PARAMETER,
-    },
+    diagnostics::codes::{Category, DiagnosticCode, DiagnosticInfo, Severity},
+    shared::files::{ByteSpan, FileByteSpan, FileId, MappedFiles},
 };
 use codespan_reporting::{
     self as csr,
@@ -36,8 +29,6 @@ use std::{
     ops::Range,
     path::PathBuf,
 };
-
-use self::codes::UnusedItem;
 
 //**************************************************************************************************
 // Types
@@ -82,28 +73,6 @@ struct JsonDiagnostic {
     category: u8,
     code: u8,
     msg: String,
-}
-
-#[derive(PartialEq, Eq, Clone, Debug)]
-/// Used to filter out diagnostics, specifically used for warning suppression
-pub struct WarningFilters {
-    filters: BTreeMap<ExternalPrefix, UnprefixedWarningFilters>,
-    for_dependency: bool, // if false, the filters are used for source code
-}
-
-#[derive(PartialEq, Eq, Clone, Debug)]
-/// Filters split by category and code
-enum UnprefixedWarningFilters {
-    /// Remove all warnings
-    All,
-    Specified {
-        /// Remove all diags of this category with optional known name
-        categories: BTreeMap<u8, Option<WellKnownFilterName>>,
-        /// Remove specific diags with optional known filter name
-        codes: BTreeMap<(u8, u8), Option<WellKnownFilterName>>,
-    },
-    /// No filter
-    Empty,
 }
 
 #[derive(PartialEq, Eq, Clone, Debug, PartialOrd, Ord)]
@@ -797,180 +766,6 @@ pub fn print_stack_trace() {
     }
 }
 
-impl WarningFilters {
-    pub const fn new_for_source() -> Self {
-        Self {
-            filters: BTreeMap::new(),
-            for_dependency: false,
-        }
-    }
-
-    pub const fn new_for_dependency() -> Self {
-        Self {
-            filters: BTreeMap::new(),
-            for_dependency: true,
-        }
-    }
-
-    pub fn is_filtered(&self, diag: &Diagnostic) -> bool {
-        self.is_filtered_by_info(&diag.info)
-    }
-
-    fn is_filtered_by_info(&self, info: &DiagnosticInfo) -> bool {
-        let prefix = info.external_prefix();
-        self.filters
-            .get(&prefix)
-            .is_some_and(|filters| filters.is_filtered_by_info(info))
-    }
-
-    pub fn union(&mut self, other: &Self) {
-        for (prefix, filters) in &other.filters {
-            self.filters
-                .entry(*prefix)
-                .or_insert_with(UnprefixedWarningFilters::new)
-                .union(filters);
-        }
-        // if there is a dependency code filter on the stack, it means we are filtering dependent
-        // code and this information must be preserved when stacking up additional filters (which
-        // involves union of the current filter with the new one)
-        self.for_dependency = self.for_dependency || other.for_dependency;
-    }
-
-    pub fn add(&mut self, filter: WarningFilter) {
-        let (prefix, category, code, name) = match filter {
-            WarningFilter::All(prefix) => {
-                self.filters.insert(prefix, UnprefixedWarningFilters::All);
-                return;
-            }
-            WarningFilter::Category {
-                prefix,
-                category,
-                name,
-            } => (prefix, category, None, name),
-            WarningFilter::Code {
-                prefix,
-                category,
-                code,
-                name,
-            } => (prefix, category, Some(code), name),
-        };
-        self.filters
-            .entry(prefix)
-            .or_insert(UnprefixedWarningFilters::Empty)
-            .add(category, code, name)
-    }
-
-    pub fn unused_warnings_filter_for_test() -> Self {
-        Self {
-            filters: BTreeMap::from([(
-                None,
-                UnprefixedWarningFilters::unused_warnings_filter_for_test(),
-            )]),
-            for_dependency: false,
-        }
-    }
-
-    pub fn for_dependency(&self) -> bool {
-        self.for_dependency
-    }
-}
-
-impl UnprefixedWarningFilters {
-    pub fn new() -> Self {
-        Self::Empty
-    }
-
-    fn is_filtered_by_info(&self, info: &DiagnosticInfo) -> bool {
-        match self {
-            Self::All => info.severity() <= Severity::Warning,
-            Self::Specified { categories, codes } => {
-                info.severity() <= Severity::Warning
-                    && (categories.contains_key(&info.category())
-                        || codes.contains_key(&(info.category(), info.code())))
-            }
-            Self::Empty => false,
-        }
-    }
-
-    pub fn union(&mut self, other: &Self) {
-        match (self, other) {
-            // if self is empty, just take the other filter
-            (s @ Self::Empty, _) => *s = other.clone(),
-            // if other is empty, or self is ALL, no change to the filter
-            (_, Self::Empty) => (),
-            (Self::All, _) => (),
-            // if other is all, self is now all
-            (s, Self::All) => *s = Self::All,
-            // category and code level union
-            (
-                Self::Specified { categories, codes },
-                Self::Specified {
-                    categories: other_categories,
-                    codes: other_codes,
-                },
-            ) => {
-                categories.extend(other_categories);
-                // remove any codes covered by the category level filter
-                codes.extend(
-                    other_codes
-                        .iter()
-                        .filter(|((category, _), _)| !categories.contains_key(category)),
-                );
-            }
-        }
-    }
-
-    /// Add a specific filter to the filter map.
-    /// If filter_code is None, then the filter applies to all codes in the filter_category.
-    fn add(
-        &mut self,
-        filter_category: u8,
-        filter_code: Option<u8>,
-        filter_name: Option<WellKnownFilterName>,
-    ) {
-        match self {
-            Self::All => (),
-            Self::Empty => {
-                *self = Self::Specified {
-                    categories: BTreeMap::new(),
-                    codes: BTreeMap::new(),
-                };
-                self.add(filter_category, filter_code, filter_name)
-            }
-            Self::Specified { categories, .. } if categories.contains_key(&filter_category) => (),
-            Self::Specified { categories, codes } => {
-                if let Some(filter_code) = filter_code {
-                    codes.insert((filter_category, filter_code), filter_name);
-                } else {
-                    categories.insert(filter_category, filter_name);
-                    codes.retain(|(category, _), _| *category != filter_category);
-                }
-            }
-        }
-    }
-
-    pub fn unused_warnings_filter_for_test() -> Self {
-        let filtered_codes = [
-            (UnusedItem::Function, FILTER_UNUSED_FUNCTION),
-            (UnusedItem::StructField, FILTER_UNUSED_STRUCT_FIELD),
-            (UnusedItem::FunTypeParam, FILTER_UNUSED_TYPE_PARAMETER),
-            (UnusedItem::Constant, FILTER_UNUSED_CONST),
-            (UnusedItem::MutReference, FILTER_UNUSED_MUT_REF),
-            (UnusedItem::MutParam, FILTER_UNUSED_MUT_PARAM),
-        ]
-        .into_iter()
-        .map(|(item, filter)| {
-            let info = item.into_info();
-            ((info.category(), info.code()), Some(filter))
-        })
-        .collect();
-        Self::Specified {
-            categories: BTreeMap::new(),
-            codes: filtered_codes,
-        }
-    }
-}
-
 impl Migration {
     pub fn new(
         mapped_files: MappedFiles,
@@ -1204,43 +999,6 @@ impl From<Vec<Diagnostic>> for Diagnostics {
 impl From<Option<Diagnostic>> for Diagnostics {
     fn from(diagnostic_opt: Option<Diagnostic>) -> Self {
         Diagnostics::from(diagnostic_opt.map_or_else(Vec::new, |diag| vec![diag]))
-    }
-}
-
-impl AstDebug for WarningFilters {
-    fn ast_debug(&self, w: &mut crate::shared::ast_debug::AstWriter) {
-        for (prefix, filters) in &self.filters {
-            let prefix_str = prefix.unwrap_or(known_attributes::DiagnosticAttribute::ALLOW);
-            match filters {
-                UnprefixedWarningFilters::All => w.write(format!(
-                    "#[{}({})]",
-                    prefix_str,
-                    WarningFilter::All(*prefix).to_str().unwrap(),
-                )),
-                UnprefixedWarningFilters::Specified { categories, codes } => {
-                    w.write(format!("#[{}(", prefix_str));
-                    let items = categories
-                        .iter()
-                        .map(|(cat, n)| WarningFilter::Category {
-                            prefix: *prefix,
-                            category: *cat,
-                            name: *n,
-                        })
-                        .chain(codes.iter().map(|((cat, code), n)| WarningFilter::Code {
-                            prefix: *prefix,
-                            category: *cat,
-                            code: *code,
-                            name: *n,
-                        }));
-                    w.list(items, ",", |w, filter| {
-                        w.write(filter.to_str().unwrap());
-                        false
-                    });
-                    w.write(")]")
-                }
-                UnprefixedWarningFilters::Empty => (),
-            }
-        }
     }
 }
