+++ external-crates/move/crates/move-compiler/src/hlir/detect_dead_code.rs
@@ -3,7 +3,10 @@
 
 use crate::{
     diag,
-    diagnostics::{Diagnostic, Diagnostics, WarningFilters},
+    diagnostics::{
+        warning_filters::{WarningFilters, WarningFiltersScope},
+        Diagnostic, Diagnostics,
+    },
     expansion::ast::ModuleIdent,
     ice,
     naming::ast::{self as N, BlockLabel},
