+++ external-crates/move/crates/move-compiler/src/naming/resolve_use_funs.rs
@@ -1,7 +1,8 @@
 // Copyright (c) The Move Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::diagnostics::{Diagnostic, Diagnostics, WarningFilters};
+use crate::diagnostics::warning_filters::{WarningFilters, WarningFiltersScope};
+use crate::diagnostics::{Diagnostic, Diagnostics};
 use crate::expansion::ast::{self as E, ModuleIdent};
 use crate::naming::ast as N;
 use crate::parser::ast::{FunctionName, Visibility};
