+++ external-crates/move/crates/move-compiler/src/naming/translate.rs
@@ -7,7 +7,8 @@ use crate::{
     diagnostics::{
         self,
         codes::{self, *},
-        Diagnostic, Diagnostics, WarningFilters,
+        warning_filters::{WarningFilters, WarningFiltersScope},
+        Diagnostic, Diagnostics,
     },
     editions::FeatureGate,
     expansion::{
