+++ external-crates/move/crates/move-compiler/src/sui_mode/info.rs
@@ -10,6 +10,7 @@ use std::{
 };
 
 use crate::{
+    diagnostics::warning_filters::WarningFilters,
     expansion::ast::{Fields, ModuleIdent},
     naming::ast as N,
     parser::ast::{Ability_, DatatypeName, Field},
@@ -271,7 +272,7 @@ fn add_private_transfers(
         transferred: &'a mut BTreeMap<(ModuleIdent, DatatypeName), TransferKind>,
     }
     impl<'a> TypingVisitorContext for TransferVisitor<'a> {
-        fn push_warning_filter_scope(&mut self, _: crate::diagnostics::WarningFilters) {
+        fn push_warning_filter_scope(&mut self, _: WarningFilters) {
             unreachable!("no warning filters in function bodies")
         }
