+++ external-crates/move/crates/move-compiler/src/typing/match_analysis.rs
@@ -3,6 +3,7 @@
 
 use crate::{
     diag,
+    diagnostics::warning_filters::WarningFilters,
     expansion::ast::{ModuleIdent, Value_},
     ice,
     naming::ast::BuiltinTypeName_,
@@ -70,7 +71,7 @@ impl TypingMutVisitorContext for MatchCompiler<'_, '_> {
         }
     }
 
-    fn push_warning_filter_scope(&mut self, filter: crate::diagnostics::WarningFilters) {
+    fn push_warning_filter_scope(&mut self, filter: WarningFilters) {
         self.context.push_warning_filter_scope(filter);
     }
