+++ external-crates/move/crates/move-stackless-bytecode/tests/testsuite.rs
@@ -5,7 +5,7 @@
 use anyhow::anyhow;
 use codespan_reporting::{diagnostic::Severity, term::termcolor::Buffer};
 use move_command_line_common::testing::EXP_EXT;
-use move_compiler::{diagnostics::WarningFilters, shared::PackagePaths};
+use move_compiler::{diagnostics::warning_filters::WarningFilters, shared::PackagePaths};
 use move_model::{model::GlobalEnv, options::ModelBuilderOptions, run_model_builder_with_options};
 use move_prover_test_utils::{baseline_test::verify_or_update_baseline, extract_test_directives};
 use move_stackless_bytecode::{
